% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylize1.R
\name{stylize1}
\alias{stylize1}
\title{Stylize Numeric Values with Custom Rules (Aiming for 1 Decimal)}
\usage{
stylize1(x)
}
\arguments{
\item{x}{Numeric value to be stylized.}
}
\value{
A character string representing the stylized value.
}
\description{
This function stylizes numeric values according to a set of custom rules.
It rounds values to 1 decimal place unless they round to a smaller digit,
and values less than 0.001 are expressed as "<0.001".
}
\examples{
stylize1(-0.05)
# Returns: "-0.1"

stylize1(0.002)
# Returns: "0.002"

stylize1(0.0005)
# Returns: "<0.001"

}
\author{
Ville Langén
}
