% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylize_p.R
\name{stylize_p}
\alias{stylize_p}
\title{Round and Format Numeric Values for Improved Readability}
\usage{
stylize_p(x)
}
\arguments{
\item{x}{Numeric value to be rounded and formatted.}
}
\value{
A character string representing the rounded and formatted value.
}
\description{
This function takes a numeric value and rounds/formats it according to specific rules
to improve readability, especially when used for p-values in statistical outputs.
The function aims to round values to 2 decimal places whenever feasible.
Specifically avoids outputting "0.05", as this value may not hold practical meaning for the reader when the significance level (alpha) is set at 0.05.
If the round result is equal to "0.050", the original value is given without any rounding.
If the original value (prior to rounding) is equal to 0.050, the original value is given without any rounding.
Values less than 0.001 are expressed as "<0.001".
}
\examples{
stylize_p(0.0499)
# Returns: "0.0499"

stylize_p(0.002)
# Returns: "0.002"

stylize_p(0.0005)
# Returns: "<0.001"

}
\author{
Ville Langén
}
