% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylize_tableone.R
\name{stylize_tableone}
\alias{stylize_tableone}
\title{Processes a tableone object to a lean, more publication-ready format}
\usage{
stylize_tableone(table1)
}
\arguments{
\item{table1}{A tableone object generated by CreateTableOne.}
}
\value{
A data frame representing the processed tableone object.
}
\description{
This function takes a tableone object and performs various operations to make
the output more suitable for publication. It includes converting rownames to
column names, transferring percentage symbols, and removing unnecessary symbols.
}
\examples{
# Assuming you already have a 'table1' object from CreateTableOne
# table1 <- CreateTableOne(data = your_data, strata = "strata_var", ...)

# Call the custom function to process the tableone object
processed_table1 <- process_tableone(table1)

# Now 'processed_table1' contains the result of all the operations

}
\seealso{
\code{\link{CreateTableOne}}
}
\author{
Ville Langen
}
\keyword{data}
\keyword{preprocessing}
\keyword{publication-ready}
\keyword{tableone}
