% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableone_1g}
\alias{tableone_1g}
\title{Example tableone Object with Imaginary Demographic Data for Practicing with the tablestylizer Package}
\format{
An example tableone object with the following components:
\describe{
\item{ContTable}{Summary statistics for continuous variables (Age (years), Height (inches)).}
\item{CatTable}{Summary statistics for categorical variables (Women, Likes coffee, Favorite color, Group).}
\item{MetaData}{Metadata information, including variable types and labels.}
}
}
\source{
This data is generated for the purpose of providing a practice example tableone object for learning and experimenting with the tablestylizer package.
It does not represent real-world individuals or situations.
}
\usage{
tableone_1g
}
\description{
This dataset is an example tableone object created using the tableone package, representing a summary table of characteristics.
It is generated based on the dataset "tablestylizer_df" and is in the format of a list with components for continuous (ContTable) and categorical (CatTable) variables.
Users can practice and explore the functionality of the tablestylizer package by tidying up this table with the tablestylizer::stylize_tableone() function.
Note that in this table, all data is presented under the "Overall" column, without dividing into Groups A, B, and C.
}
\examples{
tableone_1g

}
\keyword{datasets}
