% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableone_3g}
\alias{tableone_3g}
\title{Example tableone Object with Imaginary Demographic Data for Practicing with the tablestylizer Package}
\format{
An example tableone object with the following components:
\describe{
\item{ContTable}{Summary statistics for continuous variables (Age (years), Height (inches)) stratified by Groups A, B, and C.}
\item{CatTable}{Summary statistics for categorical variables (Women, Likes coffee, Favorite color) stratified by Groups A, B, and C.}
\item{MetaData}{Metadata information, including variable types and labels.}
}
}
\source{
This data is generated for the purpose of providing a practice tableone object with multiple strata for learning and experimenting with the tablestylizer package.
It does not represent real-world individuals or situations.
}
\usage{
tableone_3g
}
\description{
This dataset is an example tableone object created using the tableone package, representing a summary table of characteristics.
It is generated based on the dataset "tablestylizer_df" and is in the format of a list with components for continuous (ContTable) and categorical (CatTable) variables.
The table is stratified by three groups: A, B, and C. For each group, the table provides summary statistics for continuous variables (Age (years), Height (inches)) and categorical variables (Women, Likes coffee, Favorite color).
Users can practice and explore the functionality of the tablestylizer package by tidying up this table with the tablestylizer::stylize_tableone() function.
}
\examples{
tableone_3g

}
\keyword{datasets}
