% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tablestylizer_df}
\alias{tablestylizer_df}
\title{Example Dataset with Imaginary Demographic Data for Practicing with the tablestylizer Package}
\format{
A tibble with 20 rows and 7 variables:
\describe{
\item{ID}{Unique identifier for each individual.}
\item{Age (years)}{Age of the individuals in years.}
\item{Women}{Indicates whether the individual is female (Yes) or not (No).}
\item{Height (inches)}{Height of the individuals in inches.}
\item{Likes coffee}{Individuals' preference for coffee (Yes or No).}
\item{Favorite color}{Favorite color of the individuals (Red, Green, or Blue).}
\item{Group}{Group affiliation of the individuals (A, B, or C).}
}
}
\source{
This data is generated for the purpose of illustrating data manipulation and analysis in R.
It does not represent real-world individuals or situations.
}
\usage{
tablestylizer_df
}
\description{
This dataset provides an example of fictional data related to a group of individuals.
It includes information such as age, sex, height, coffee preferences, favorite color, and group affiliation.
}
\examples{
head(tablestylizer_df)

}
\keyword{datasets}
