% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylize1.R
\name{stylize1}
\alias{stylize1}
\title{Stylize Numeric Values with Custom Rules (Aiming for 1 Decimal)}
\usage{
stylize1(x, threshold = NULL)
}
\arguments{
\item{x}{Numeric value to be stylized.}
}
\value{
A character string representing the stylized value.
}
\description{
Stylizes numeric values according to a set of custom rules.
Rounds values to 1 decimal place unless they round to a smaller digit.
Optional parameter to specify a threshold for displaying values as "<threshold".
Very large or small values are expressed as powers of 10.
}
\examples{
stylize1(-0.05)
# Returns: "-0.1"

stylize1(0.002)
# Returns: "0.002"

stylize1(0.0005)
# Returns: "<0.001"

}
\author{
Ville Langén
}
