% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylize2.R
\name{stylize2}
\alias{stylize2}
\title{Stylize Numeric Values with Custom Rules (Aiming for 2 Decimals)}
\usage{
stylize2(x)
}
\arguments{
\item{x}{Numeric value to be stylized.}
}
\value{
A character string representing the stylized value.
}
\description{
This function stylizes numeric values according to a set of custom rules.
It rounds values to 2 decimal places unless they round to a smaller digit,
and values less than 0.001 are expressed as "<0.001".
}
\examples{
stylize2(-0.05)
# Returns: "-0.05"

stylize2(0.002)
# Returns: "0.00"

stylize2(0.0005)
# Returns: "<0.001"

}
\author{
Ville Langén
}
